requirements.txt
pymongo>=4.0.0
requests>=2.25.0
python-dotenv>=0.19.0
loguru>=0.7.3
